+++ packages/core/src/index.ts
@@ -1,6 +1,7 @@
 export * from "./abi";
 export * from "./addresses";
 export * from "./fees";
+export * from "./flow";
 export * from "./names";
 export * from "./position";
 export * from "./program";